requirements.txt
fpdf
pymupdf
openpyxl
pandas
fpdf
jinja2
streamlit
numpy
tavily-python
pymongo
#nltk
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy

# AI dependencies
openai
crewai
ollama
langchain
transformers==4.46.0
tokenizers==0.20.3
sentence-transformers == 3.4.1
google-genai
chromadb
crewai-tools




